% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse-functions.R
\name{confidenceEllipse}
\alias{confidenceEllipse}
\title{Bivariate Normal Confidence Ellipse}
\usage{
confidenceEllipse(
  X.mean = c(0, 0),
  eig,
  n,
  p,
  xl = NULL,
  yl = NULL,
  axes = TRUE,
  center = FALSE,
  lim.adj = 0.02,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{X.mean}{a column matrix giving the mean of the two dimensions of the
p-dimensional multivariate normal distribution.}

\item{eig}{the eigenvalues and eigenvectors of the covariance matrix. This
should be of the same form as the output of \code{\link{eigen}}, namely a
list with two components: \code{values} and \code{vectors}. It is assumed
that the largest eigenvalue is given first.}

\item{n}{the number of observations.}

\item{p}{the number of dimensions of the multivariate normal distribution.
(The resulting graph will always be a two-dimensional confidence region for
the two dimensions of a p-dimensional multivaraite normal distribution under
consideration.)}

\item{xl}{a vector giving the lower and upper limits of the x-axis for
plotting. If \code{xl = NULL} (default), then reasonable values are computed
automatically.}

\item{yl}{a vector giving the lower and upper limits of the y-axis for
plotting. If \code{yl = NULL} (default), then reasonable values are computed
automatically.}

\item{axes}{logical. If \code{axes = TRUE} (default) then the major and
minor axes of the ellipse are plotted.}

\item{center}{logical. If \code{center = TRUE} then the center of the ellipse
is indicated with a point and dashed lines are drawn to the x-axis and
y-axis.}

\item{lim.adj}{a value giving an adjustment to the x-axis and y-axis limits
computed if either \code{xl = NULL} or \code{yl = NULL}. Essentially this is
a way to have some coarse control over these limits for quick graphing:
positive values will increase the distance between the upper and lower
limits (making the ellipse appear smaller) while negative values will
decrease the distance (and make the ellipse appear larger).}

\item{alpha}{a value giving the value of alpha to be used when computing the
contour. Contours are drawn at the \code{1-alpha} level.}

\item{...}{other arguments to be passed to the graphing functions.}
}
\value{
None
}
\description{
Draws a (1-\code{alpha})100\% confidence ellipse (two dimensional) for a
multivariate normal distribution using the eigendecomposition of the
covariance matrix.
}
\examples{

# 90\% Confidence Ellipse for Reading and Vocab from ability.cov
x.bar <- ability.cov$center[5:6]
Sigma <- ability.cov$cov[5:6,5:6]
n <- ability.cov$n.obs
p <- length(ability.cov$center)

confidenceEllipse(X.mean = x.bar,
                  eig = eigen(Sigma),
                  n = n, p = p,
                  alpha = 0.10)

}
\references{
Johnson, R. A., & Wichern, D. W. (2007). Applied multivariate
statistical analysis (6th ed). Pearson Prentice Hall.
}
