% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse-functions.R
\name{eigenEllipseHelper}
\alias{eigenEllipseHelper}
\title{Helper Function for other
Ellipse-from-Eigendecomposition Functions}
\usage{
eigenEllipseHelper(mu, lengths, angle, xl, yl, lim.adj, axes, center, ...)
}
\arguments{
\item{mu}{column matrix giving the coordinates for the cener of the ellipse.}

\item{lengths}{vector giving the major and minor axis lengths.}

\item{angle}{angle of rotation (in radians).}

\item{xl}{x-axis limits. If \code{xl = NULL} then these are computed
automatically.}

\item{yl}{y-axis limits. If \code{yl = NULL} then these are computed
automatically.}

\item{lim.adj}{a value giving an adjustment to the x-axis and y-axis limits
computed if either \code{xl = NULL} or \code{yl = NULL}.}

\item{axes}{logical. If \code{axes = TRUE}, then the major and minor axes are
graphed.}

\item{center}{logical. If \code{center = TRUE} then the center of the ellipse
is indicated with a point and dashed lines are drawn to the x-axis and
y-axis.}

\item{\dots}{other arguments to be passed to the graphing functions.}
}
\value{
None
}
\description{
Helper function for graphing ellipses from eigendecompositions. This function
is used by \code{\link{bvNormalContour}} and \code{\link{confidenceEllipse}}.
Essentially this is a wrapper for \code{\link[plotrix]{draw.ellipse}} that
also calculates appropriate x-axis and y-axis limits to make graphing an
ellipse easier (because the entire ellipse should be visible without any
work on the user's part to specify the limits).
}
