% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot-rectangles.R
\name{make_rect}
\alias{make_rect}
\title{Add a rectangle to a correlation plot}
\usage{
make_rect(
  rstart,
  rend,
  cstart = NULL,
  cend = NULL,
  correlation_matrix,
  mirror = FALSE,
  lwd = 3,
  ...
)
}
\arguments{
\item{rstart}{The name of the variable contained in the correlation plot that will serve as the vertical starting point of the rectangle.}

\item{rend}{(optional) The name of the variable contained in the correlation plot that will serve as the horizontal ending point of the rectangle.
If no cstart or cend provided, creates an on-diagonal rectangle automatically. cstart and cend allows fo off-diagonal rectangles.}

\item{cstart}{(optional) The name of the variable contained in the correlation plot that will serve as the horizontal starting point of the rectangle.
If no cstart or cend provided, creates an on-diagonal rectangle automatically. cstart and cend allows fo off-diagonal rectangles.}

\item{correlation_matrix}{The correlation matrix the correlation plot is based on.}

\item{mirror}{If TRUE, also adds the equivalent rectangle from the other side of the diagonal (i.e. the "mirror" of the original).}

\item{...}{Arguments to modify graphical parameters, etc.}
}
\description{
Add a rectangle to a correlation plot
}
\examples{
#Adding an on-diagonal rectangle to a correlation plot of the USArrests data with mirroring.
corrplot(cor(USArrests))
make_rect(rstart="Assault", rend="UrbanPop", correlation_matrix=cor(USArrests), mirror=TRUE)

#Adding an off-diagonal rectangle
make_rect(rstart="Murder", rend="Rape", cstart="Murder", cend="Murder", correlation_matrix=cor(USArrests))

}
